pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-community",
    "pypdf",
    "pymupdf",
    "langchain-text-splitters",
    "sentence-transformers",
    "faiss-cpu",
    "chromadb",
    "langchain-groq",
    "python-dotenv",
    "typesense",
    "langchain-openai",
    "langgraph",
    "langchain-nomic",
    "tiktoken",
    "langchainhub",
    "tavily-python",
    "gpt4all",
    "fastembed",
    "beautifulsoup4",
    "lxml",
    "fastapi",
    "uvicorn",
    "pydantic",
    "unicorn>=2.1.4",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.21",
    "google-search-results>=2.4.2",
    "serpapi>=0.1.5",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.11",
    "python-jose>=3.5.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "scrapy>=2.14.1",
    "selenium>=4.6",
    "scrapy-selenium>=0.0.7",
    "webdriver-manager>=4.0.2",
    "torch>=2.9.1",
]
